Implemented run and threads.
3 threads. Initial idea behind them:
1) MANAGER thread: Do the simulating task without worrying with printing and the time.
2) PRINTER thread: Do the printing task without worrying with the simulating and the time.
3) TIMER thread:   Do the timining task (i.e. write new/update time) without worrying with printing and simulating.

In order to make sure they will work properly, there will be protected variables (using semaphores).
IDEA:
- Semaphore A to protect memory files and the free lists (MUTEX Manager versus Printer)
- Semaphore B to protect 'now' time (MUTED Printer versus Timer) (*)

a) Everytime a new interval of time is completed, the printer thread prints memory
the whole files and free lists once it acquires pass permission from Semaphore A

(*) Still must see if this semaphore can be avoided (if there's a way to assign atomically the time).